fix: Use int32_t for disk struct fields to ensure 4-byte size on 64-bit (#387)
* fix: Use int32_t for disk struct fields to ensure 4-byte size on 64-bit

The tydisktreenode, tydisktreerec, and tyOLD42disktreenode structs use
`long` for fields like nodevalsize, ctnodes, and flags. On 64-bit
platforms (LP64 model), `long` is 8 bytes, but the disk format expects
these fields to be 4 bytes.

This caused assertion failures:
  assert(sizeof(tydisktreenode) == 12)  // was 16-18 bytes on 64-bit
  assert(sizeof(tydisktreerec) == 18)   // was wrong size

Changed `long` to `int32_t` (from <stdint.h>) for:
- tydisktreenode.nodevalsize
- tydisktreerec.ctnodes
- tydisktreerec.flags
- tyOLD42disktreenode.nodevalsize

Now the structs are correctly sized regardless of platform word size.

Fixes #386


* fix: Remove duplicate #include <stdint.h>

Address PR review feedback - stdint.h was already included at line 36.


---------

+++ Common/source/langtree.c
@@ -53,7 +53,7 @@ typedef struct tydisktreenode {
 	tytreetype nodetype; /*add, subtract, if, etc.*/
 #endif
 
-	long nodevalsize;
+	int32_t nodevalsize;  /* Must be 4 bytes on all platforms for disk format */
 	
 	short lnum; /*which line number in the source was this node generated from?*/
 	
@@ -100,15 +100,15 @@ typedef enum tydisktreenodeparaminfo {
 
 #pragma pack(2)
 typedef struct tydisktreerec {
-	
+
 	short version;
-	
-	long ctnodes;
-	
-	long flags; /*currently unused*/
-	
+
+	int32_t ctnodes;  /* Must be 4 bytes on all platforms for disk format */
+
+	int32_t flags; /*currently unused - must be 4 bytes for disk format*/
+
 	byte waste [8];
-	
+
 	tydisktreenode nodes [];
 	} tydisktreerec, *ptrdisktreerec, **hdldisktreerec;
 
@@ -122,7 +122,7 @@ typedef struct tyOLD42disktreenode {
 	short nodetype;
 	#endif
 
-	long nodevalsize;
+	int32_t nodevalsize;  /* Must be 4 bytes on all platforms for disk format */
 	
 	short lnum; /*which line number in the source was this node generated from?*/
